requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
ecdsa==0.19.0
idna==3.7
josepy==1.14.0
mozilla-django-oidc==4.0.1
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1

PyJWT~=2.8.0